backend/wrrapd-api-repo/WrrapdServer/python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.12.0
requests==2.31.0
Pillow==10.2.0
gunicorn==21.2.0

